+++ src/ui/bookmarks_tab.rs
@@ -26,7 +26,7 @@ use crate::env::JjConfig;
 use crate::env::get_env;
 use crate::ui::Component;
 use crate::ui::ComponentAction;
-use crate::ui::help_popup::HelpPopup;
+use crate::ui::dialog::HelpPopup;
 use crate::ui::message_popup::MessagePopup;
 use crate::ui::panel::DetailsPanel;
 use crate::ui::panel::TextContent;
