+++ src/ui/dialog/mod.rs
@@ -9,6 +9,8 @@ until it is closed.
 
 mod bookmark_set;
 mod command;
+mod help;
 
 pub use bookmark_set::BookmarkSetPopup;
 pub use command::CommandPopup;
+pub use help::HelpPopup;
