+++ src/ui/files_tab.rs
@@ -20,7 +20,7 @@ use crate::env::JjConfig;
 use crate::env::get_env;
 use crate::ui::Component;
 use crate::ui::ComponentAction;
-use crate::ui::help_popup::HelpPopup;
+use crate::ui::dialog::HelpPopup;
 use crate::ui::message_popup::MessagePopup;
 use crate::ui::panel::DetailsPanel;
 use crate::ui::panel::TextContent;
